Most of my focus has been on the mergeSort function so far. Still trying different things but I've realized that I haven't made any commits to the repo yet.
I'd welcome anyone else to try running the code to see if they get different results but my "optimization" is consistently about an order of magnitude slower than the unoptimized code.

Been trying several things with the parallel pragma, however I think I'm just creating so many threads that it's bogging things down... Might try to do a conditional pragma based on the size of the section to be sorted?